Rename websocketprocessor to remoteobserverprocessor
