+++ cmd/otelcontribcol/processors_test.go
@@ -23,9 +23,9 @@ import (
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/attraction"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/attributesprocessor"
+	remoteobserverprocessor "github.com/open-telemetry/opentelemetry-collector-contrib/processor/remoteobserverprocessor"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/resourceprocessor"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/spanprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/websocketprocessor"
 )
 
 func TestDefaultProcessors(t *testing.T) {
@@ -138,9 +138,9 @@ func TestDefaultProcessors(t *testing.T) {
 			processor: "redaction",
 		},
 		{
-			processor: "websocket",
+			processor: "remoteobserver",
 			getConfigFn: func() component.Config {
-				cfg := procFactories["websocket"].CreateDefaultConfig().(*websocketprocessor.Config)
+				cfg := procFactories["remoteobserver"].CreateDefaultConfig().(*remoteobserverprocessor.Config)
 				cfg.Endpoint = "localhost:0"
 				return cfg
 			},
