+++ processor/remoteobserverprocessor/channelset_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package websocketprocessor
+package remoteobserverprocessor
 
 import (
 	"testing"
