+++ processor/remoteobserverprocessor/internal/metadata/generated_status.go
@@ -7,7 +7,7 @@ import (
 )
 
 const (
-	Type             = "websocket"
+	Type             = "remoteobserver"
 	LogsStability    = component.StabilityLevelDevelopment
 	MetricsStability = component.StabilityLevelDevelopment
 	TracesStability  = component.StabilityLevelDevelopment
