+++ processor/remoteobserverprocessor/metadata.yaml
@@ -1,4 +1,4 @@
-type: websocket
+type: remoteobserver
 
 status:
   class: processor
